pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "archon-rh"
version = "0.1.0"
description = "Offline reasoning lab for formal mathematics research on the Riemann Hypothesis."
authors = [{name = "ARCHON Team"}]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8.1.7,<9.0",
  "numpy>=1.26,<2.0",
  "pandas>=2.1,<3.0",
  "pyarrow>=14.0,<17.0",
  "torch==2.2.2",
  "torchaudio==2.2.2",
  "torchvision==0.17.2",
  "accelerate>=0.31,<0.33",
  "transformers>=4.40,<4.46",
  "sentencepiece>=0.1.99,<0.2.0",
  "bitsandbytes>=0.43,<0.44; platform_system == 'Linux'",
  "ray[default]==2.10.0",
  "networkx>=3.2,<3.4",
  "mpmath>=1.3,<2.0",
  "sympy>=1.12,<1.13",
  "rich>=13.7,<14.0",
  "pydantic>=2.7,<2.8",
  "pyyaml>=6.0,<7.0",
  "polars>=0.20,<0.21",
  "fastparquet>=2024.2,<2025.0",
  "pyarrow-hotfix",
  "ruff>=0.4.2,<0.5.0",
  "pytest>=8.2,<8.3",
  "pytest-asyncio>=0.23,<0.24",
  "jsonschema>=4.21,<4.22",
  "tqdm>=4.66,<5.0",
  "datasets>=2.19,<2.20",
  "huggingface-hub>=0.22,<0.25",
  "tenacity>=8.2,<9.0",
  "uvicorn>=0.29,<0.30",
  "fastapi>=0.111,<0.112",
  "typer>=0.12,<0.13",
  "xxhash>=3.4,<4.0",
  "orjson>=3.10,<3.11",
  "cryptography>=42,<43",
  "ed25519>=1.5,<1.7",
  "shortuuid>=1.0,<2.0"
]

[project.optional-dependencies]
dev = [
  "ruff",
  "pytest",
  "pytest-cov",
  "mypy",
  "types-PyYAML",
  "types-requests"
]

[tool.setuptools]
packages = {find = {include = ["formal*", "prover*", "conjecture*", "numerics*", "orchestration*", "safety*"]}}

[tool.ruff]
select = ["E", "F", "I", "UP", "N"]
line-length = 100

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["formal/tests", "prover/tests", "conjecture/tests", "numerics/tests", "safety/tests"]
pythonpath = ["."]

[tool.mypy]
exclude = ["formal/leanlib"]
python_version = "3.10"
warn_unused_ignores = true
warn_return_any = true
